LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-after-discard.https-expected.txt


PASS getUserMedia() in a discarded browsing context
FAIL enumerateDevices() in a discarded browsing context assert_equals: Promise state expected "pending" but got "resolved"
